optimize: optimize synchronization procedure (using requeue instead of goroutine) and reconstruct project layout
